finalized encryptData/decryptData in BlockCipherAbstract

+++ src/symmetric/cipher/stream/php_stream_cipher_abstract.cpp
@@ -317,7 +317,7 @@ PHP_METHOD(Cryptopp_StreamCipherAbstract, encryptData) {
     }
 
     // check dataSize against block size
-    long blockSize = static_cast<long>(encryptor->MandatoryBlockSize());
+    int blockSize = static_cast<int>(encryptor->MandatoryBlockSize());
 
     if (0 != dataSize % blockSize) {
         zend_class_entry *ce;
@@ -353,7 +353,7 @@ PHP_METHOD(Cryptopp_StreamCipherAbstract, decryptData) {
     }
 
     // check dataSize against block size
-    long blockSize = static_cast<long>(decryptor->MandatoryBlockSize());
+    int blockSize = static_cast<int>(decryptor->MandatoryBlockSize());
 
     if (0 != dataSize % blockSize) {
         zend_class_entry *ce;
